join all dict funcs into utils
